Python 3 compatibility issue when no verbose arguments specified fixed
In Python 2, `None > 0` is a valid comparison that returns `False`. In
Python 3, it is not, and raises an error. Previously, if no "-v" or
"--verbose" arguments were specified, then `verbose` would default to
None. Now it defaults properly to 0.

Also added .gitignore for compiled Python files and Pycharm .idea
folder.